[JENKINS-36905] make "cleanSlashes" a shared util
